# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "싱송생송 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/keep": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "플레이리스트에 있는 노래들을 가져온다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Playlist"
                ],
                "summary": "플레이리스트에 노래를 가져온다",
                "responses": {
                    "200": {
                        "description": "성공",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/handler.PlaylistAddResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "노래들을 하나씩 플레이리스트에 추가한 후 적용된 플레이리스트의 노래들을 리턴한다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Playlist"
                ],
                "summary": "플레이리스트에 노래를 추가한다",
                "parameters": [
                    {
                        "description": "노래 리스트",
                        "name": "PlaylistAddRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PlaylistAddRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/handler.PlaylistAddResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "노래들을 하나씩 플레이리스트에서 삭제한다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Playlist"
                ],
                "summary": "플레이리스트에 노래를 제거한다",
                "parameters": [
                    {
                        "description": "노래 리스트",
                        "name": "SongDeleteFromPlaylistRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SongDeleteFromPlaylistRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.PlaylistAddResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/v2/keep": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "플레이리스트에 있는 노래들을 가나다순/최신추가순/오래된순(alphabet/recent/old) 으로 가져온다. 기본값은 recent이다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Playlist"
                ],
                "summary": "플레이리스트에 노래를 여러가지 필터로 가져온다 (커서기반 페이징)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "필터",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "마지막에 조회했던 커서의 songId(이전 요청에서 lastCursor값을 주면 됨), 없다면 default로 가장 최신 글부터 조회",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "한번에 조회할 노래의 개수. 입력하지 않는다면 기본값인 20개씩 조회",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.GetPlayListV2Response"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.GetPlayListV2Response": {
            "type": "object",
            "properties": {
                "lastCursor": {
                    "type": "integer"
                },
                "songs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.PlaylistAddResponse"
                    }
                }
            }
        },
...
        "handler.PlaylistAddRequest": {
            "type": "object",
            "properties": {
                "songId": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "handler.PlaylistAddResponse": {
            "type": "object",
            "properties": {
                "album": {
                    "type": "string"
                },
                "isLive": {
                    "type": "boolean"
                },
                "isMr": {
                    "type": "boolean"
                },
                "keepSongId": {
                    "type": "integer"
                },
                "lyricsYoutubeLink": {
                    "type": "string"
                },
                "melonLink": {
                    "type": "string"
                },
                "singerName": {
                    "type": "string"
                },
                "songId": {
                    "type": "integer"
                },
                "songName": {
                    "type": "string"
                },
                "songNumber": {
                    "type": "integer"
                },
                "tjYoutubeLink": {
                    "type": "string"
                }
            }
        },
...
        "handler.SongDeleteFromPlaylistRequest": {
            "type": "object",
            "properties": {
                "songIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "pkg.BaseResponseStruct": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}